requirements.txt
bs4
requests
ebooklib
pyyaml
